refine external SFT rollout path and cleanup tests
