+++ docs/on_policy_distillation.md
@@ -62,27 +62,46 @@ enabled = false  # Skip expensive verification
 
 This runs pure on-policy distillation: the student learns to match the teacher without needing any reward signal.
 
-## SFT Distillation From Teacher Rollouts (Text-Only)
+## SFT Distillation ("Hard Distillation") From Teacher Rollouts
 
-When the teacher endpoint only returns text (no token IDs/logprobs), you can still use the RL orchestrator/trainer stack with an SFT objective:
+Use this mode when you want to train from teacher-generated completions directly (hard distillation), without teacher token-level logprobs.
 
 ```toml
 [trainer.loss]
 type = "sft"
 
-[orchestrator.rollout_model.client]
+[orchestrator]
+use_token_client = false
+
+[orchestrator.teacher_rollout_model.client]
 base_url = ["https://your-openai-compatible-endpoint/v1"]
 skip_model_check = true
 
-[orchestrator.rollout_model.model]
+[orchestrator.teacher_rollout_model.model]
 name = "teacher-model-name"
 ```
 
 In this mode:
-- Rollouts are generated from `orchestrator.rollout_model`
-- The orchestrator automatically uses text-level reconstruction with the student tokenizer (`use_token_client=false`)
+- Rollouts are generated from `orchestrator.teacher_rollout_model`
+- The orchestrator uses text-level reconstruction with the student tokenizer
 - The RL trainer optimizes masked NLL (`trainer.loss.type = "sft"`)
 - Omit `[inference]` (no local inference server required)
+
+### Image Input (VLM) Support
+
+Yes, image input is supported in SFT/hard-distillation mode when the student model is multimodal (VLM).
+
+- Prompts can include OpenAI-style image items in `message.content`, e.g. `{"type": "image_url", "image_url": {"url": "data:image/..."}}`
+- The orchestrator extracts and preprocesses images from trajectory prompts and attaches `pixel_values`/`image_grid_thw` to training samples
+- No teacher token IDs/logprobs are required; reconstruction still happens from messages
+
+Notes:
+- This path currently expects `data:image/...` payloads in message content
+- The teacher rollout endpoint still needs to be able to handle the same multimodal prompts during generation
+
+Reference configs:
+- `configs/alphabet_sort/sft_distill_hard_qwen4b_lora_prime_teacher.toml`
+- `examples/alphabet_sort/sft_distill_hard.toml`
 
 ## Parameters
