+++ examples/alphabet_sort/sft_distill_hard.toml
@@ -1,4 +1,4 @@
-max_steps = 11
+max_steps = 24
 seq_len = 2048
 
 [deployment]
@@ -10,6 +10,10 @@ num_infer_gpus = 0
 [model]
 name = "Qwen/Qwen3-4B-Instruct-2507"
 
+[wandb]
+project = "alphabet-sort-sft-distill"
+name = "qwen3-4b-hard-distill"
+
 [ckpt]
 
 [trainer.optim]
@@ -19,17 +23,19 @@ lr = 1e-5
 type = "sft"
 
 [trainer.model.lora]
+rank = 16
+alpha = 32
 
 [trainer.ckpt.weights]
 save_adapter_separately = true
 
 [orchestrator]
-batch_size = 128
-rollouts_per_example = 2
+batch_size = 256
+rollouts_per_example = 4
 use_token_client = false
 
 [orchestrator.sampling]
-max_tokens = 384
+max_tokens = 512
 temperature = 0.7
 
 [orchestrator.teacher_rollout_model.client]
